# ui/contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DKNet.Accounts.Api",
    "version": "1",
    "description": "DRK-1684 §3 row 1 — committed snapshot of the ledger service's own published OpenAPI document. Regenerated and diffed by `verify:contract` (row 3); never hand-edited."
  },
  "servers": [{ "url": "/v1" }],
  "paths": {
...
    "/postings": {
      "get": {
        "operationId": "listPostings",
        "summary": "List postings across every account, within a required effective-date window of at most 90 days.",
        "security": [{ "bearer": ["postings.read"] }],
        "parameters": [
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "accountId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "direction", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "orderBy", "in": "query", "schema": { "type": "string" } },
          { "name": "desc", "in": "query", "schema": { "type": "boolean" } },
          { "name": "pageNumber", "in": "query", "schema": { "type": "integer" } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PagedPostingResponse" } } } },
          "400": { "description": "Refused", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefusalBody" } } } }
        }
      },
      "post": {
        "operationId": "recordPosting",
        "summary": "Record one credit or debit. Idempotency key is required in the header.",
        "security": [{ "bearer": ["postings.write"] }],
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecordPostingRequest" } } }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PostingDto" } } } },
          "422": { "description": "Refused", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefusalBody" } } } },
          "409": { "description": "Idempotency key conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefusalBody" } } } }
        }
      }
    },
    "/postings/{id}": {
      "get": {
        "operationId": "getPosting",
        "summary": "Read one posting.",
        "security": [{ "bearer": ["postings.read"] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PostingDto" } } } },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/postings/{id}/reverse": {
      "post": {
        "operationId": "reversePosting",
        "summary": "Reverse a posting. Requires a reason in the body and an idempotency key in the header.",
        "security": [{ "bearer": ["postings.reverse"] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["reason"], "properties": { "reason": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PostingDto" } } } },
          "422": { "description": "Refused", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefusalBody" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "PostingDto": {
        "type": "object",
        "required": ["id", "postingNumber", "accountId", "streamPosition", "direction", "amount", "signedAmount", "currency", "status"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "postingNumber": { "type": "string" },
          "accountId": { "type": "string", "format": "uuid" },
          "streamPosition": { "type": "integer" },
          "direction": { "type": "string" },
          "amount": { "type": "string" },
          "signedAmount": { "type": "string" },
          "currency": { "type": "string" },
          "status": { "type": "string" },
          "category": { "type": "string" },
          "description": { "type": "string" },
          "effectiveDate": { "type": "string", "format": "date" },
          "reversesPostingId": { "type": "string", "format": "uuid" },
          "reversedByPostingId": { "type": "string", "format": "uuid" }
        }
      },
      "PagedPostingResponse": {
        "type": "object",
        "required": ["items", "pageIndex", "pageSize", "pageCount", "hasNextPage"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/PostingDto" } },
          "pageIndex": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "pageCount": { "type": "integer" },
          "hasNextPage": { "type": "boolean" }
        }
      },
      "RecordPostingRequest": {
        "type": "object",
        "required": ["accountId", "direction", "amount", "currency", "category"],
        "properties": {
          "accountId": { "type": "string", "format": "uuid" },
          "direction": { "type": "string", "enum": ["Debit", "Credit"] },
          "amount": { "type": "string" },
          "currency": { "type": "string" },
          "category": { "type": "string" },
          "description": { "type": "string" },
          "effectiveDate": { "type": "string", "format": "date" }
        }
      },
...
      "ErrorItem": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" },
          "code": { "type": "string" },
          "field": { "type": "string" }
        }
      },
      "RefusalBody": {
        "type": "object",
        "required": ["status", "errors", "traceId"],
        "properties": {
          "status": { "type": "integer" },
          "errors": { "type": "array", "items": { "$ref": "#/components/schemas/ErrorItem" } },
          "traceId": { "type": "string" }
        }
      }
    }
  }
}